Fix one way collision checking for getting valid depths on margins for KinematicBody2D
Makes sure collisions are consistent for the margins in all collision checks
with the test_body_motion function. The collision solver also
has the valid direction passed in for one way collisions in order
to get the proper best axis (i.e. not a perpindicular axis).

There are changes the one way collision check in the callback.
The reposition check now allows angles below 45 degrees
and above 0 degrees to be counted for one way collision,
allowing for steeper angles to still work for one way collision.

Finally, valid depths are no longer changed by either body's velocity.
Instead, the collision body's velocity is taken into account during
the actual collision solving methods. Also, during scaling of the
body's velocity for repositioning, if the collision is a one way
collision, another attempt is made to change the scale so that
the body lands within the margins.